storybook: move over ui-core stories
