+++ storybook/stories/ui-core/RadioButton.stories.tsx
@@ -1,8 +1,7 @@
 import { type StoryObj, type Meta } from '@storybook/react';
 
+import { RadioButton } from '@osrd-project/ui-core';
 import '@osrd-project/ui-core/dist/theme.css';
-
-import RadioButton from '../components/inputs/RadioButton';
 
 const meta: Meta<typeof RadioButton> = {
   component: RadioButton,
