+++ storybook/stories/ui-core/Select.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { type StoryObj, type Meta } from '@storybook/react';
 
-import Select, { type SelectProps } from '../components/Select';
+import { Select, type SelectProps } from '@osrd-project/ui-core';
 import '@osrd-project/ui-core/dist/theme.css';
 
 type Option = { value: string; label: string };
